requirements.txt
# Core Flask dependencies
Flask==2.3.3
Werkzeug==2.3.7

# Database
pymongo==4.6.0
dnspython==2.4.2

# Machine Learning
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2

# AI and Image Processing
google-generativeai==0.3.2
Pillow==10.1.0

# Environment and Configuration
python-dotenv==1.0.0

# Weather API (OpenWeatherMap)
requests==2.31.0
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0